uppsrc: Fixed many Win64 warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@5821 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Sql/Sqlexp.h
@@ -191,8 +191,8 @@ public:
 	SqlS(const char *s, int pr)             : text(s), priority(pr) {}
 	SqlS(const String& s, int pr)           : text(s), priority(pr) {}
 	
-	force_inline SqlS(const SqlS& a, const char *op, const SqlS& b, int pr, int prb) { Init(a, op, strlen(op), b, pr, prb); }
-	force_inline SqlS(const SqlS& a, const char *op, const SqlS& b, int pr)          { Init(a, op, strlen(op), b, pr, pr); }
+	force_inline SqlS(const SqlS& a, const char *op, const SqlS& b, int pr, int prb) { Init(a, op, (int)strlen(op), b, pr, prb); }
+	force_inline SqlS(const SqlS& a, const char *op, const SqlS& b, int pr)          { Init(a, op, (int)strlen(op), b, pr, pr); }
 };
 
 class SqlVal : public SqlS, Moveable<SqlVal> {
